Break out synchronization code into its own module, to avoid circular imports when using util.py.
